Move the Rig Creator to be inline with Bgeo creator
